+++ Task14/app/src/main/java/com/example/will/task14/OutputFragment.java
@@ -1,10 +1,6 @@
 package com.example.will.task14;
 
-import android.content.Context;
-import android.net.Uri;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
 import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -12,7 +8,7 @@ import android.view.ViewGroup;
 import android.widget.TextView;
 
 
-public class Output extends Fragment {
+public class OutputFragment extends Fragment {
     private TextView textView;
 
     @Override
@@ -22,14 +18,4 @@ public class Output extends Fragment {
         return inflater.inflate(R.layout.fragment_output, container, false);
     }
 
-    @Override
-    public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
-        super.onViewCreated(view, savedInstanceState);
-
-        this.textView = (TextView) view.findViewById(R.id.textView);
-    }
-
-    void setTextView(String text){
-        this.textView.setText(text);
-    }
 }
