+++ Task14/app/src/main/res/layout/fragment_output.xml
@@ -3,13 +3,13 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Output">
+    tools:context=".OutputFragment">
 
     <!-- TODO: Update blank fragment layout -->
 
     <TextView
         android:id="@+id/textView"
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:hint="ここに表示されます" />
+        android:layout_height="match_parent"
+        android:hint="@string/willOutput" />
 </FrameLayout>
