+++ Task14/app/src/main/res/values/strings.xml
@@ -1,8 +1,8 @@
 <resources>
     <string name="app_name">Task14</string>
-    <string name="hint">ボタンを押すとここに表示されます</string>
-    <string name="button">Button</string>
 
     <!-- TODO: Remove or change this placeholder text -->
-    <string name="hello_blank_fragment">Hello blank fragment</string>
+    <string name="pleaseInput">入力してください</string>
+    <string name="button">Button</string>
+    <string name="willOutput">ここに表示されます</string>
 </resources>
